feat(webgpu): report the whole adapter, honestly
The adapter reply carried an id and a name; AdapterInfo has eight fields
and DeviceCaps behind one of them. It now carries all of it, so what a
browser reports reaches the capability seam rather than a summary of it.
The reply format is version 2, because the two halves ship as separately
cached artifacts and a stale decoder has to fail at the header rather
than read a name's length as a vendor id.

The fields WebGPU has no answer for are marked unknown rather than
filled in: no numeric vendor or device id, no discrete-versus-integrated,
no driver string. A fabricated id would be indistinguishable downstream
from a real one, and isFallbackAdapter grades performance rather than
device class, so it is not that answer either.

The feature mapping is explicit and lossy both ways, and both lists are
written where the mapping lives. The browser gate compares the feature
word wasm received against one the page computes from the live adapter,
with the bit numbers spelled out in the driver instead of imported from
the code under test.

One bug found on the way: a reply was written field by field and a
failure mid-write left half an adapter in the buffer, which refuses the
whole frame's replies. Replies are written atomically now.

+++ crates/crcbl-webgpu/src/bytes.rs
@@ -74,8 +74,13 @@ pub enum DecodeError {
     InvalidEnum {
         /// Field that carried it.
         field: &'static str,
-        /// The code that was read.
-        code: u32,
+        /// The code that was read, widened to the largest a field can carry.
+        ///
+        /// A `u64` because [`Features`](crcbl_hal::Features) is a 64-bit
+        /// bitflags and is refused through `from_bits` rather than truncated —
+        /// a narrower field here would have to drop the half of the value that
+        /// says which bit was unclaimed.
+        code: u64,
     },
     /// A handle field that may not be `None` held zero bits.
     #[error("{field} is a null handle")]
